AnimPicker: Decompose gui.py and picker_widgets.py; extract PickerItemData (Phase 2)
Behavior-preserving decomposition of the two anim_picker God-files into focused
modules, on top of the Phase 1 cleanup. Public API, importable names, Shifter's
MAYA_OVERRIDE_COLOR, and the .pkr / PICKER_DATAS formats are unchanged.

- gui.py (2361 -> 40 lines, shim) split into constants.py, scene.py, view.py,
  tab_widget.py, main_window.py.
- widgets/picker_widgets.py (2942 -> 53 lines, shim) split into widgets/graphics.py,
  widgets/picker_item.py, and widgets/dialogs/ (script_dialog, search_replace_dialog,
  handles_window, item_options, copy_paste_dialog).
- New Qt/Maya-free widgets/item_model.py (PickerItemData); PickerItem.get_data /
  set_data delegate to its to_dict / from_dict, preserving the exact schema round-trip.
- gui.py and picker_widgets.py kept as compatibility shims re-exporting the same
  objects (isinstance identity preserved). Import cycles broken with lazy imports
  (copy_paste_dialog -> PickerItem, tab_widget -> MainDockWindow).
- Curve<->picker conversion kept in view.py; picker_curves.py extraction deferred.
